athalscan_requirements.txt
argparse
Pillow
numpy
python-math
regex
pandas

